use mediumint for most stat columns to save 25% storage size for stat tables and another 60% for referrer_urls table

+++ migrations/1.0-initial-schema.php
@@ -8,8 +8,8 @@ $wpdb->query( "DROP TABLE IF EXISTS {$wpdb->prefix}koko_analytics_site_stats" );
 $wpdb->query(
 	"CREATE TABLE {$wpdb->prefix}koko_analytics_site_stats (
 		   date DATE PRIMARY KEY NOT NULL,
-		   visitors INTEGER UNSIGNED NOT NULL,
-		   pageviews INTEGER UNSIGNED NOT NULL
+		   visitors MEDIUMINT UNSIGNED NOT NULL,
+		   pageviews MEDIUMINT UNSIGNED NOT NULL
 	) ENGINE=INNODB CHARACTER SET=ascii"
 );
 
@@ -18,8 +18,8 @@ $wpdb->query(
 	"CREATE TABLE {$wpdb->prefix}koko_analytics_post_stats (
 	   date DATE NOT NULL,
 	   id BIGINT(20) UNSIGNED NOT NULL,
-	   visitors INTEGER UNSIGNED NOT NULL,
-	   pageviews INTEGER UNSIGNED NOT NULL,
+	   visitors MEDIUMINT UNSIGNED NOT NULL,
+	   pageviews MEDIUMINT UNSIGNED NOT NULL,
 	   PRIMARY KEY (date, id)
 	) ENGINE=INNODB CHARACTER SET=ascii"
 );
@@ -28,9 +28,9 @@ $wpdb->query( "DROP TABLE IF EXISTS {$wpdb->prefix}koko_analytics_referrer_stats
 $wpdb->query(
 	"CREATE TABLE {$wpdb->prefix}koko_analytics_referrer_stats (
 	   date DATE NOT NULL,
-	   id BIGINT(20) UNSIGNED NOT NULL,
-	   visitors INTEGER UNSIGNED NOT NULL,
-	   pageviews INTEGER UNSIGNED NOT NULL,
+	   id MEDIUMINT UNSIGNED NOT NULL,
+	   visitors MEDIUMINT UNSIGNED NOT NULL,
+	   pageviews MEDIUMINT UNSIGNED NOT NULL,
 	   PRIMARY KEY (date, id)
 	) ENGINE=INNODB CHARACTER SET=ascii"
 );
@@ -38,7 +38,7 @@ $wpdb->query(
 $wpdb->query( "DROP TABLE IF EXISTS {$wpdb->prefix}koko_analytics_referrer_urls" );
 $wpdb->query(
 	"CREATE TABLE {$wpdb->prefix}koko_analytics_referrer_urls (
-	   id BIGINT(20) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
+	   id MEDIUMINT UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
 	   url VARCHAR(255) NOT NULL,
 	   UNIQUE INDEX (url)
 	) ENGINE=INNODB CHARACTER SET=ascii"
